requirements.txt
mastodon.py
dotenv
os
json
openai
tweepy
argparse
shutil
datetime
requests
beautifulesoup4
